requirements.txt
django~=4.1
djangorestframework==3.15.1
PyMySQL==1.1.1
mysqlclient==2.2.4
djangorestframework-simplejwt==5.3.1
django-allauth==0.63.3
